+++ .eslintrc
@@ -14,6 +14,7 @@
         "extensions": [".js"]
       }
     ],
+    "react/jsx-props-no-spreading" :0,
     "react/state-in-constructor": [0],
     "no-param-reassign": 0,
     "jsx-a11y/label-has-associated-control": [
