+++ package.json
@@ -21,7 +21,8 @@
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "react-scripts build && echo '/* /index.html 200' | cat >build/_redirects"
+    "build": "react-scripts build && echo '/* /index.html 200' | cat >build/_redirects",
+    "test": "react-scripts test"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -39,6 +40,7 @@
     ]
   },
   "devDependencies": {
+    "@testing-library/react": "^11.0.4",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.8.0",
     "eslint-config-airbnb": "^18.2.0",
@@ -46,10 +48,12 @@
     "eslint-plugin-import": "^2.22.0",
     "eslint-plugin-jsx-a11y": "^6.3.1",
     "eslint-plugin-react": "^7.21.1",
-    "eslint-plugin-react-hooks": "^2.5.1",
+    "eslint-plugin-react-hooks": "^4.1.2",
     "husky": "^4.3.0",
+    "jest-styled-components": "^7.0.3",
     "lint-staged": "^10.4.0",
-    "prettier": "2.0.4"
+    "prettier": "2.1.2",
+    "react-test-renderer": "^16.13.1"
   },
   "husky": {
     "hooks": {
