+++ src/components/atoms/Fieldset/Fieldset.js
@@ -1,24 +1,12 @@
 ﻿import React from 'react';
-import styled from 'styled-components';
 import PropTypes from 'prop-types';
-
-const StyledWrapper = styled.fieldset`
-  border-radius: ${({ theme }) => theme.borderRadius};
-  color: white;
-  padding: 25px;
-
-  legend {
-    font-size: ${({ theme }) => theme.fontSize.normal};
-    padding: 0 10px;
-    margin-left: 15px;
-  }
-`;
+import { Wrapper } from './styled/StyledFieldset';
 
 const Fieldset = ({ legend, children, settings }) => (
-  <StyledWrapper settings={settings}>
+  <Wrapper settings={settings}>
     <legend>{legend}</legend>
     {children}
-  </StyledWrapper>
+  </Wrapper>
 );
 
 Fieldset.defaultProps = {
