+++ src/components/atoms/index.js
@@ -1,12 +1,12 @@
-﻿import Button from './Button';
-import Title from './Title';
-import Fieldset from './Fieldset';
-import Input from './Input';
-import Logo from './Logo';
-import Paragraph from './Paragraph';
-import ButtonIcon from './ButtonIcon';
-import Select from './Select';
-import SubTitle from './SubTitle';
+﻿import Button from './Button/Button';
+import Title from './Title/Title';
+import Fieldset from './Fieldset/Fieldset';
+import Input from './Input/Input';
+import Logo from './Logo/Logo';
+import Paragraph from './Paragraph/Paragraph';
+import ButtonIcon from './ButtonIcon/ButtonIcon';
+import Select from './Select/Select';
+import SubTitle from './SubTitle/SubTitle';
 
 export {
   Button,
