+++ src/components/molecules/Basket.js
@@ -1,6 +1,6 @@
 ﻿import React from 'react';
 import styled from 'styled-components';
-import Title from '../atoms/Title';
+import Title from '../atoms/Title/Title';
 
 const StyledWrapper = styled.div`
   background-color: white;
